Blockchain/requirement.txt
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==7.0.0
pycparser==2.19
requests==2.22.0
six==1.14.0
urllib3==1.25.8
Werkzeug==0.16.0
WTForms==2.2.1
